==> Files/2_data_wrangling.ipynb <==
## I.2 Explore and perform Data Wrangling on:

1. Dataset: Scrapped Dataset of Properties listed in Birmingham B28 + 1 mile Radius (rightmove.co.uk)
2. Scrapped date: 4, Nov, 2024
3. Author: Lam H. Tong

# Import packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import geopy

# Import scrapped dataset

# All properties listed
df_all     = pd.read_csv('[all]B28_plus_1mile_rightmove_4_Nov.csv')

# Properties with garden only
df_garden = pd.read_csv('[garden]B28_plus_1mile_rightmove_4_Nov.csv')

# Properties with parking only
df_parking = pd.read_csv('[parking]B28_plus_1mile_rightmove_4_Nov.csv')

# New home only
df_new    = pd.read_csv('[new]B28_plus_1mile_rightmove_4_Nov.csv')

# The main dataframe info before wrangling
df_all.info()

# remove excessive characters
for df in [df_garden, df_new, df_parking, df_all]:
    df['price'] = df['price'].str.replace('£', '')

# drop duplicates (as every scrapped pages have 1 same advertised property):
df_all = df_all.drop_duplicates()
df_garden = df_garden.drop_duplicates()
df_new = df_new.drop_duplicates()
df_parking = df_parking.drop_duplicates()

# Adding featured coulumns in the filtered dataset 
df_parking['parking'] = True
df_garden['garden'] = True
df_new['new_home'] = True

#change 'no_bed' type to string to perform merging bc sh only this feild of this particular dataset is int64
df_garden['no_bed'] = df_garden['no_bed'].astype(str)

# merge the df_all with df_garden first
df_temp1 = pd.merge(df_all, df_garden, on = ['address', 'price', 'type', 'no_bed', 'no_bath', 'agent', 'url'] , how = 'outer')

# merge with df_parking
df_temp2 = pd.merge(df_temp1, df_parking, on = ['address', 'price', 'type', 'no_bed', 'no_bath', 'agent', 'url'] , how = 'outer')

#change 'no_bed' type to string to perform merging
df_new['no_bed'] = df_new['no_bed'].astype(str)

# the final merged dataframe
df = pd.merge(df_temp2, df_new, on = ['address', 'price', 'type', 'no_bed', 'no_bath', 'agent', 'url'] , how = 'outer')

# Define the regex pattern for UK postcodes
postcode_pattern = r'(\bB\d{1,2}(?:\s?\d[A-Z]{2})?\b)'

# Extract postcodes using .str.extract() and the pattern if there're any
df['postcode'] = df['address'].str.extract(postcode_pattern)

# Replace excessive newlines and carriage returns with a single space
df['address'] = df['address'].str.replace(r'[\n\r]+', ' ', regex=True)

# Replace 'not_specified' with NaN for `no_bed` and `no_bath`
df['no_bed'].replace('not_specified', np.nan, inplace=True)
df['no_bath'].replace('not_specified', np.nan, inplace=True)

# Remove commas and convert `price` to int
df['price'] = pd.to_numeric(df['price'].str.replace(',', ''), errors='coerce')

# Convert `no_bed` and `no_bath` to numeric, allowing NaNs
df['no_bed'] = pd.to_numeric(df['no_bed'], errors='coerce')
df['no_bath'] = pd.to_numeric(df['no_bath'], errors='coerce')

# Convert to nullable integer type (Int64) to handle NaNs
df['no_bed'] = df['no_bed'].astype('Int64')  # Nullable int
df['no_bath'] = df['no_bath'].astype('Int64')  # Nullable int

# Replacing Nulls in 'garden', 'parking', and 'new_home' as False
for feature in ['garden', 'parking', 'new_home']:
    df.loc[df[feature].isna(), feature] = False

df[df['no_bath'].isna()]

# Correcting Bathrooms manunally as per the provided floor plan AND description of each property
for idx in [0, 79, 82, 220, 303, 639, 640, 655]:
    df.loc[idx, 'no_bath'] = 1

for idx in [71]:
    df.loc[idx, 'no_bath'] = 2

for idx in [433, 494, 495, 496, 497, 498, 499, 500, 501, 502, 503, 504, 505, 506, 507, 508, 509, 510, 548]:
    df.loc[idx, 'no_bath'] = 3

df[df['no_bed'].isna()]

#Remove the above properties as they are commercial buildings
df = df.drop(index = [219, 628])

# Correcting types of these properties
df.loc[58, 'type'] = 'Flat'
df.loc[88, 'type'] = 'Semi-Detached'

# Drop the Commercial Property 
df = df.drop(686, axis = 0)

# Export to a separted dadaset
df.to_csv('B28_properties_dataset_cleaned.csv', index=False)
