==> Files/3_data_preparation.ipynb <==
## I.3 INITAL DATA PREPARATION

The complete dataset of the Population: Properties in B28 Area (Birmingham - UK)

# import packages
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
from geopy.geocoders import Nominatim
from geopy.extra.rate_limiter import RateLimiter

# Import the main dataset
df = pd.read_csv('B28_properties_dataset_cleaned.csv')

# Inspect the dataframe
df.head()

# Dataframe info
df.info()

- Population size:   834
- Inital attributes: 10

To-do(s):
1. Get a sample of 120 properties
2. Get the location (longtitude, lattitude) of these properties for geospatial visualisations
3. Scrapping the nearest Station/School -> Go back to add properties URL in the scrapping step

### 1. Get a sample of 120 by simple sampling:

# Get a sample of 120 using Simple Random
df_sample = df.sample(n =120, random_state = 42)

### 2. Get the location (longtitude, lattitude) of these properties for geospatial visualisations

geolocator = Nominatim(user_agent="property_locator")
geocode = RateLimiter(geolocator.geocode, min_delay_seconds=1)

# Apply geocoding to each row with error handling for null or unfound addresses
df_sample['location'] = df_sample['address'].apply(lambda x: geocode(x) if pd.notnull(x) and x.strip() != "" else None)
df_sample['latitude'] = df_sample['location'].apply(lambda loc: loc.latitude if loc else None)
df_sample['longitude'] = df_sample['location'].apply(lambda loc: loc.longitude if loc else None)

# Properties than can't be decoded by geolocator
df_sample[df_sample['location'].isna()]

def adding_location(row, location, latitude, longitude):
    """
    This function corrects the location and coordinates of properties by updating null cells
    with the provided location, latitude, and longitude values.

    Arguments:
        row -- index of the row to update
        location -- new location/address to add if missing
        latitude -- latitude coordinate to add if missing
        longitude -- longitude coordinate to add if missing
    """
    if pd.isna(df_sample.loc[row, 'location']):
        df_sample.loc[row, 'location'] = location
    if pd.isna(df_sample.loc[row, 'latitude']):
        df_sample.loc[row, 'latitude'] = latitude
    if pd.isna(df_sample.loc[row, 'longitude']):
        df_sample.loc[row, 'longitude'] = longitude


# mannually adding the locations and coordinates of 12 properties whose location cound't be found using google map:
adding_location(30, 'Ashdale Drive, Maypole, Hollywood, Birmingham, West Midlands, England, B14, United Kingdom', 52.402974437344255, -1.8791687366426464)
adding_location(599, 'Shaftmoor Lane, Hall Green, Birmingham, West Midlands, England, B28 8SW, United Kingdom', 52.442364101124284, -1.8427314180797243)
adding_location(596, 'Shaftmoor Lane, Hall Green, Birmingham, West Midlands, England, B28 8SW, United Kingdom', 52.442364101124284, -1.8427314180797243)
adding_location(76, '356 Wake Green Road, Moseley, Birmingham, West Midlands, England, B13, United Kingdom', 52.431594828004734, -1.8572746283589694)
adding_location(155, 'Croome Close, Sparkhill, Birmingham, West Midlands, England, B11, United Kingdom', 52.44882446111353, -1.870650026528147)
adding_location(412, 'Park Road, Birmingham, West Midlands, England, B11, United Kingdom', 52.44853865010483, -1.868405114984758)
adding_location(622, '53 Wake Green Road, Moseley, Birmingham, West Midlands, England, B13 9HW, United Kingdom', 52.44380463429303, -1.878170617197836)
adding_location(331, 'Loxley Court, Baldwins Lane, Birmingham, West Midlands, England, B28 0FH,	United Kingdom', 52.42095047459805, -1.8383903929671765)
adding_location(324, 'Littlemead Road, Shirley, Solihull, Birmingham, West Midlands, England, B90 , United Kingdom', 2.39688028331134, -1.8501598271600033)

df_sample['location']

# Export the Sampled Dataset with location information for later analysis
df.to_csv('[120sample]B28_properties.csv]')

### 3. Get a separate dataset for testing the Trained Model:

# Get a sample of 30 using Simple Random
df_test = df.sample(n =30, random_state = 23)

df_test.head()

df_test.to_csv('[Test]B28_30_Properties.csv')
